requirements.txt
flask
pandas
python-dotenv
openai                 # optional; used if you plug OpenAI
